rename module from communicator to communication
it is the place for communication related things, so communication
is a more general and on-topic name
