+++ test/unit/communication/test_communicator_grid.cpp
@@ -8,7 +8,7 @@
 // SPDX-License-Identifier: BSD-3-Clause
 //
 
-#include "dlaf/communicator/communicator_grid.h"
+#include "dlaf/communication/communicator_grid.h"
 
 #include <mpi.h>
 #include <gtest/gtest.h>
